move registration to eos-vm.cpp
